refactor: move chapter mapping to ChapterFrom
