+++ backend/graph/api/episodes.resolvers.go
@@ -223,65 +223,13 @@ func (r *episodeResolver) Chapters(ctx context.Context, obj *model.Episode) ([]*
 	metadataItems = lo.Filter(metadataItems, func(i *common.TimedMetadata, _ int) bool {
 		return i.Type == "chapter"
 	})
-	ginCtx, _ := utils.GinCtx(ctx)
-	languages := user.GetLanguagesFromCtx(ginCtx)
 
 	r.Loaders.PhraseLoader.LoadMany(ctx, lo.Uniq(lo.Map(metadataItems, func(i *common.TimedMetadata, _ int) string {
 		return i.ChapterType.Value
 	})))
 
 	return lo.Map(metadataItems, func(i *common.TimedMetadata, _ int) *model.Chapter {
-		title := i.Title.Get(languages)
-		phrase, _ := r.Loaders.PhraseLoader.Get(ctx, i.ChapterType.Value)
-		emptyTitle := title == ""
-		if emptyTitle && phrase != nil {
-			title = phrase.Value.Get(languages)
-		}
-
-		switch i.ChapterType {
-		case common.ChapterTypeSong, common.ChapterTypeSingAlong:
-			if !i.SongID.Valid {
-				break
-			}
-			song, _ := r.Loaders.SongLoader.Get(ctx, i.SongID.UUID)
-			if song == nil {
-				break
-			}
-			if emptyTitle {
-				if phrase != nil {
-					title = fmt.Sprintf("%s - %s", phrase.Value.Get(languages), song.Title.Get(languages))
-				} else {
-					title = song.Title.Get(languages)
-				}
-			} else {
-				title = strings.Replace(title, "{{song.title}}", song.Title.Get(languages), -1)
-			}
-		case common.ChapterTypeSpeech, common.ChapterTypeAppeal, common.ChapterTypeTestimony:
-			if len(i.PersonIDs) != 1 {
-				break
-			}
-			personID := i.PersonIDs[0]
-			person, _ := r.Loaders.PersonLoader.Get(ctx, personID)
-			if person == nil {
-				break
-			}
-			if emptyTitle {
-				if phrase != nil {
-					title = fmt.Sprintf("%s - %s", phrase.Value.Get(languages), person.Name)
-				} else {
-					title = person.Name
-				}
-			} else {
-				title = strings.Replace(title, "{{person.name}}", person.Name, -1)
-			}
-		}
-
-		return &model.Chapter{
-			ID:          i.ID.String(),
-			Title:       title,
-			Description: i.Description.GetValueOrNil(languages),
-			Start:       int(i.Timestamp),
-		}
+		return model.ChapterFrom(ctx, i, r.Loaders)
 	}), nil
 }
